backend/requirements.txt
Django
djangorestframework
django-cors-headers
requests
reportlab
gunicorn
whitenoise